+++ tt_metal/hw/ckernels/blackhole/metal/llk_api/experimental/llk_unpack_fast_tilize_api.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "llk_unpack_tilize_api.h"
+#include "experimental/llk_unpack_fast_tilize.h"
 
 /*************************************************************************
  * LLK UNPACK FAST TILIZE SRC A (BH)
